requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
tortoise-tts==3.0.0
numpy>=1.20.0
pydub>=0.25.1
transformers==4.31.0
einops>=0.6.1
rotary_embedding_torch>=0.2.0
unidecode>=1.3.6
scipy>=1.10.0
librosa>=0.10.0 